change for gprs

+++ Classes/keypadview.h
@@ -22,10 +22,10 @@
 	UIImage *keypadImageP;
 	UIImage *pressedImageP;
 	int touchDown;
-	int elementx;
-	int elementy;
-	int elementWidth;
-	int elementHeight;
+	float elementx;
+	float elementy;
+	float elementWidth;
+	float elementHeight;
 	CGRect rectchange;
 	id<KeypadProtocol> keypadProtocolP;
